[Design] report speaking rate chart fixed

+++ highpitch/Presentation/MainWindow/ProjectView/PracticeView/Store/PracticeViewStore.swift
@@ -134,17 +134,17 @@ extension PracticeViewStore {
     }
     
     /// (문장 중앙 시간, SPM)
-    func getSentenceSpeakingRateData() -> [(Int, Double)] {
-        var answer: [(Int, Double)] = []
-        answer.append((0, practice.summary.spmAverage))
+    func getSentenceSpeakingRateData() -> [(Double, Double)] {
+        var answer: [(Double, Double)] = []
+        answer.append((0.0, practice.summary.spmAverage))
         for sentence in getSortedSentences() {
             answer.append((
-                Int(Double(sentence.startAt + sentence.endAt) / 2000.0),
+                Double(sentence.startAt + sentence.endAt) / 2000.0,
                 sentence.spmValue
             ))
         }
         answer.append((
-            Int(practice.summary.practiceLength),
+            practice.summary.practiceLength,
             practice.summary.spmAverage
         ))
         return answer
